Controlador de puntos pero con chequeo de permisos para cada usuario

See merge request proyecto2021/proyectos/grupo40!24

+++ app/resources/punto.py
@@ -6,6 +6,7 @@ from app.db import db
 from app.resources.validadorPuntos import ValidarForm
 
 from app.helpers.auth import authenticated
+from app.helpers.auth import hasPermission
 from app.models.punto import Punto
 from app.models.configuracion import Configuracion
 
@@ -15,6 +16,10 @@ from app.models.configuracion import Configuracion
 def index():
     if not authenticated(session):
        abort(401)
+
+    if not hasPermission("punto_encuentro_index", session.get("id")):
+        abort(401)
+
     conf=Configuracion.getConfigs()
     params=request.args
     currentPage = int(params.get("page", 0))
@@ -25,6 +30,10 @@ def index():
 def create():
     if not authenticated(session):
        abort(401)
+
+    if not hasPermission("punto_encuentro_new", session.get("id")):
+        abort(401)
+
     params=request.form
     mensaje=ValidarForm(params)
     if mensaje.validate()==False:
@@ -45,6 +54,9 @@ def create():
 def update(id):
 
     if not authenticated(session):
+        abort(401)
+
+    if not hasPermission("punto_encuentro_update", session.get("id")):
         abort(401)
 
     params=request.form
@@ -81,6 +93,9 @@ def delete(id):
     if not authenticated(session):# or not admin(session):
        abort(401)
 
+    if not hasPermission("punto_encuentro_destroy", session.get("id")):
+        abort(401)
+
     punto_to_delete=Punto.query.get_or_404(id)
     try:
         db.session.delete(punto_to_delete)
@@ -92,6 +107,9 @@ def delete(id):
 def show(id):
 
     if not authenticated(session): #or not admin(session):
+        abort(401)
+
+    if not hasPermission("punto_encuentro_show", session.get("id")):
         abort(401)
 
     p=Punto.query.get_or_404(id)
